rename config.rs
